Completed verification tests & IO choosing
In total, there are 6 tests that must not fail in order to create a custom circuit.

After this is verified, the user is able to choose inputs and outputs in the order that they would like generated. While this has been functionally implemented, there are also plans to have a label (optional) for each custom circuit.